Octomap plugin parameters in yaml file.

+++ srs_env_model/include/srs_env_model/but_server/registration/pcl_registration_module.h
@@ -150,7 +150,7 @@ protected:
 	// SCA features
 
 	//! Minimum distances between samples
-	float m_scaMinSampleDistance;
+	double m_scaMinSampleDistance;
 
 	//! Number of samples to use during each iteration.
 	int m_scaNumOfSamples;
